[TASK] Initial demo written in PHP added
The demo as it stands now shows how to calculate the signature required
to generate the request to the API endpoint.

It uses version 3 of the API and generates the request to list all route
names and numbers for all routes of all route types ("/v3/route").

The user ID ("devid") and API key (128bit GUID developer key) need to be
stored in a configuration file ".env". See "env.sample" for details.